+++ lib/mongoid-minitest.rb
@@ -15,6 +15,7 @@ require "matchers/validations/format"
 require "matchers/validations/inclusion"
 require "matchers/validations/exclusion"
 require "matchers/validations/confirmation"
+require "matchers/validations/acceptance"
 require "matchers/associations/associations"
 
 module Mongoid