+++ test/matchers/validations_test.rb
@@ -8,6 +8,7 @@ describe "Validations" do
     it { must validate_presence_of(:role).with_message("no role") }
 
     it { must validate_confirmation_of(:password) }
+    it { must validate_acceptance_of(:terms_of_use) }
 
     it { must validate_uniqueness_of(:login).case_insensitive }
     it { must validate_uniqueness_of(:login).scoped_to(:site) }