+++ test/matchers/associations_test.rb
@@ -7,6 +7,7 @@ describe "Associations" do
 
     it { must have_many(:pets).of_type(Pet) }
     it { must have_one(:account).of_type(Account) }
+    it { must have_and_belong_to_many(:friends).of_type(Person) }
   end
 
   describe Pet do